requirements.txt
flask
scikit-learn
openai
python-dotenv
numpy
sentence-transformers
faiss-cpu
pypdf
tqdm
